# faster viz to_program [pr]