`VIZ=1 python3 -m pytest ./test/test_multitensor.py`

```
rewrote 2737 graphs and matched 315223 times, saved to /var/folders/vw/qw2gc7rs53jdykwhx85mcxf00000gn/T/rewrites.pkl.qazal
*** viz is starting
preparing profile: 100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 9172/9172 [00:00<00:00, 1154957.76it/s]
*** started viz on http://127.0.0.1:8000
*** ready in 2433.36ms
```

master:

```
rewrote 2737 graphs and matched 315460 times, saved to /var/folders/vw/qw2gc7rs53jdykwhx85mcxf00000gn/T/rewrites.pkl.qazal
*** viz is starting
preparing profile: 100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 9182/9182 [00:00<00:00, 1034164.37it/s]
*** started viz on http://127.0.0.1:8000
*** ready in 7391.52ms
```
pcall is just dumb, codegen errors should be handled in render_program call, ie. `FUSE_ARANGE=1 VIZ=1 IGNORE_OOB=0 CPU=1 python test/test_ops.py TestOps.test_scatter_add`

![image](https://github.com/user-attachments/assets/add6e53d-a288-4f83-bd2d-6dc140836be7)

UOp.__repr__ does not need to be protected. If a UOp cannot be serialized/deserialized it should fail earlier in the pickle.load.